refactor: make Lead RAG opt-in per agent
