+++ src/Domains/Intelligence/Agents/Neuron/BaseKanvasAgent.php
@@ -14,19 +14,14 @@ use Kanvas\Intelligence\Agents\Models\Agent;
 use Kanvas\Intelligence\Agents\Neuron\Tools\Common\CurrentTimeTool;
 use Kanvas\Intelligence\Agents\Neuron\Tools\DynamicSubAgentTool;
 use Kanvas\Intelligence\Agents\Services\AgentProviderService;
-use Kanvas\Intelligence\Knowledge\Retrieval\LeadKnowledgeRetrieval;
-use Kanvas\Intelligence\Knowledge\Services\LeadRagComponents;
 use Kanvas\Intelligence\Sessions\Models\Session;
 use Kanvas\NervousSystem\Capability\Models\Tool;
 use Kanvas\Users\Models\Users;
 use NeuronAI\Agent\SystemPrompt;
 use NeuronAI\Providers\AIProviderInterface;
-use NeuronAI\RAG\Embeddings\EmbeddingsProviderInterface;
 use NeuronAI\RAG\RAG;
-use NeuronAI\RAG\Retrieval\RetrievalInterface;
-use NeuronAI\RAG\VectorStore\MemoryVectorStore;
-use NeuronAI\RAG\VectorStore\VectorStoreInterface;
 use NeuronAI\Tools\ToolInterface;
+use NeuronAI\Workflow\Node;
 use Override;
 
 class BaseKanvasAgent extends RAG implements ProvidesToolDependencies
@@ -175,35 +170,17 @@ class BaseKanvasAgent extends RAG implements ProvidesToolDependencies
     }
 
     /**
-     * Neuron's native RAG pipeline calls this retrieval before it augments the prompt.
-     * It becomes a no-op when the app flag is disabled or the turn has no current Lead.
+     * RAG is opt-in. A concrete agent enables it by using a trait that provides
+     * leadRagNodes(), such as HasLeadRag.
+     *
+     * @return list<Node>
      */
     #[Override]
-    protected function retrieval(): RetrievalInterface
+    protected function ragNodes(): array
     {
-        return new LeadKnowledgeRetrieval($this->resolveLeadForTurn());
-    }
-
-    #[Override]
-    protected function embeddings(): EmbeddingsProviderInterface
-    {
-        if ($this->app === null) {
-            throw new ValidationException(
-                'App not set. Call setConfiguration() before resolving RAG embeddings.'
-            );
-        }
-
-        return LeadRagComponents::embeddings($this->app);
-    }
-
-    #[Override]
-    protected function vectorStore(): VectorStoreInterface
-    {
-        $lead = $this->resolveLeadForTurn();
-
-        return $lead !== null
-            ? LeadRagComponents::vectorStore($lead)
-            : new MemoryVectorStore();
+        return method_exists($this, 'leadRagNodes')
+            ? $this->leadRagNodes()
+            : [];
     }
 
     /**
